backend/requirements.txt
fastapi
python-multipart
uvicorn
azure-ai-projects
azure-identity
python-dotenv